ConfluenceGroupManager/requirements.txt
protobuf
google-api-python-client
googleapis-common-protos
gevent
grpcio
grpcio-tools
requests
pyopenssl
Pillow
bs4